fix negative ports

+++ extensions/arcnet/src/arc/net/dns/SRVRecord.java
@@ -1,13 +1,13 @@
 package arc.net.dns;
 
 public final class SRVRecord implements Comparable<SRVRecord>{
-    public final int ttl;
-    public final short priority;
-    public final short weight;
+    public final long ttl;
+    public final int priority;
+    public final int weight;
     public final int port;
     public final String target;
 
-    public SRVRecord(int ttl, short priority, short weight, int port, String target){
+    public SRVRecord(long ttl, int priority, int weight, int port, String target){
         this.ttl = ttl;
         this.priority = priority;
         this.weight = weight;
@@ -18,9 +18,9 @@ public final class SRVRecord implements Comparable<SRVRecord>{
     @Override
     public int compareTo(SRVRecord o){
         if(this.priority != o.priority){
-            return Short.compare(this.priority, o.priority);
+            return Integer.compare(this.priority, o.priority);
         }else{
-            return Short.compare(this.weight, o.weight);
+            return Integer.compare(this.weight, o.weight);
         }
     }
 }
